fix_all_add_Makefile

+++ ft_substr.c
@@ -1,4 +1,5 @@
 #include <stddef.h>
+#include <stdlib.h>
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
@@ -11,7 +12,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	if (!result)
 		return (NULL);
 	i = 0;
-	while (i < (int)len && s[start + i] != '\0')
+	while (i < len && s[start + i] != '\0')
 	{
 		result[i] = s[start + i];
 		i++;
